requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
anthropic==0.40.0
bcrypt==4.1.3
PyJWT==2.9.0
pydantic[email]==2.9.2
httpx==0.27.2
